Refactor: extracted Sentiment subjectivity code into its own module
