+++ tests/high_level/test_sentiment_subjectivity.py
@@ -4,7 +4,7 @@ import numpy as np
 import pytest
 
 from nlp_profiler.constants import NOT_APPLICABLE, NaN
-from nlp_profiler.core import sentiment_subjectivity_score, \
+from nlp_profiler.sentiment_subjectivity import sentiment_subjectivity_score, \
     sentiment_subjectivity_summarised, sentiment_subjectivity  # noqa
 
 objective_subjective_text = '2833047 and 1111 people live in this area.'
